Add ObjectManager and SimWrapper for scene object management and rendering
